python/ch07/langchain/langgraph_basic_agent/requirements.txt
langchain
langchain-openai
python-dotenv
langgraph
